move recombee creds to instance settings
