+++ packages/lesswrong/lib/publicSettings.ts
@@ -99,9 +99,6 @@ export const mapboxAPIKeySetting = new DatabasePublicSetting<string | null>('map
 export const mailchimpForumDigestListIdSetting = new DatabasePublicSetting<string | null>('mailchimp.forumDigestListId', null)
 export const mailchimpEAForumListIdSetting = new DatabasePublicSetting<string | null>('mailchimp.eaForumListId', null)
 
-export const recombeeDatabaseIdSetting = new DatabasePublicSetting<string | null>('recombee.databaseId', null);
-export const recombeePublicApiTokenSetting = new DatabasePublicSetting<string | null>('recombee.publicApiToken', null);
-
 export const isProductionDBSetting = new DatabasePublicSetting<boolean>('isProductionDB', false)
 
 export const showReviewOnFrontPageIfActive = new DatabasePublicSetting<boolean>('annualReview.showReviewOnFrontPageIfActive', true)
